warnings: int --> size_t
Former-commit-id: 82da1bf6ff538a7e65b9d738c87fb3caf6e25bff

+++ modules/DataLayer/Storage/Helpers.h
@@ -127,8 +127,8 @@ namespace VELaSSCo
   // returns NULL if dst_len is too short, otherwise return dst
   inline const char *ToHexString( char *dst, size_t dst_len, const char *src, const size_t src_len, const char *format="%02x") {
     if ( !dst) return NULL;
-    int isrc = 0;
-    for ( int idst = 0; 
+    size_t isrc = 0;
+    for ( size_t idst = 0; 
 	  ( isrc < src_len) && ( idst < dst_len - 1); 
 	  isrc++, idst += 2) {
       sprintf( &dst[ idst], format, ( unsigned char)src[ isrc]);
@@ -140,8 +140,8 @@ namespace VELaSSCo
   inline const char *FromHexString( char *dst, size_t dst_len, const char *src, const size_t src_len) {
     if ( !dst) return NULL;
     bool error = false;
-    int isrc = 0;
-    for ( int idst = 0; 
+    size_t isrc = 0;
+    for ( size_t idst = 0; 
 	  ( isrc < src_len) && ( idst < dst_len); 
 	  isrc += 2, idst++) {
       int hi4 = tolower( src[ isrc]);
